requirements.txt
Flask
Gunicorn
line-bot-sdk
openai
